+++ app/Http/Controllers/Admin/UsersController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Admin;
 
 use App\Handlers\Curl;
 use App\Models\Coordinate;
+use App\Models\Department;
 use App\Models\User;
 use App\Models\UserCategory;
 use Illuminate\Http\Request;
@@ -20,15 +21,16 @@ class UsersController extends Controller
         return view('admin.user.index', compact('users'));
     }
 
-    public function create(User $user, UserCategory $userCategory)
+    public function create(User $user, UserCategory $userCategory, Department $department)
     {
         $categories = $userCategory->all();
-        return view('admin.user.create_and_edit', compact('user', 'categories'));
+        $departments = $department->all();
+        return view('admin.user.create_and_edit', compact('user', 'categories', 'departments'));
     }
 
     public function store(Request $request, User $user)
     {
-        $data = $request->only(['name', 'phone', 'password', 'age',  'category_id']);
+        $data = $request->only(['name', 'phone', 'password', 'age',  'category_id', 'department_id']);
         $data['password'] = Hash::make($data['password']);
         $user->fill($data);
         $user->save();
@@ -39,16 +41,17 @@ class UsersController extends Controller
         ]);
     }
 
-    public function edit(User $user, UserCategory $userCategory)
+    public function edit(User $user, UserCategory $userCategory, Department $department)
     {
         $categories = $userCategory->all();
-        return view('admin.user.create_and_edit', compact('user', 'categories'));
+        $departments = $department->all();
+        return view('admin.user.create_and_edit', compact('user', 'categories', 'departments'));
     }
 
     public function update(Request $request, User $user)
     {
 
-        $data = $request->only(['name', 'phone', 'password', 'age',  'category_id']);
+        $data = $request->only(['name', 'phone', 'password', 'age',  'category_id', 'department_id']);
         if ($user->password != $data['password']) {
             $data['password'] = Hash::make($data['password']);
         }
@@ -62,6 +65,12 @@ class UsersController extends Controller
 
     public function destroy(User $user)
     {
+        if ($user->event()->count()>0 || $user->dailyTask()->count()>0 || $user->commonTask()->count()>0) {
+            return response()->json([
+                'status' => 400,
+                'message' => '请删除当前用户下的所有数据'
+            ]);
+        }
         $user->delete();
 
         return response()->json([
